Refactor features to ease the addition of new ones
Defines an interface for all features and maintain an array of features
to update the game state regardless of how many there are.
