+++ src/features/experience.h
@@ -1,0 +1,8 @@
+#ifndef __FEATURE_EXPERIENCE_H__
+#define __FEATURE_EXPERIENCE_H__
+
+#include "../feature_management.h"
+
+struct Feat init_feat_experience();
+
+#endif
