+++ src/features/health.h
@@ -1,0 +1,8 @@
+#ifndef __FEATURE_HEALTH_H__
+#define __FEATURE_HEALTH_H__
+
+#include "../feature_management.h"
+
+struct Feat init_feat_health();
+
+#endif
