+++ src/features/hunger.h
@@ -1,0 +1,8 @@
+#ifndef __FEATURE_HUNGER_H__
+#define __FEATURE_HUNGER_H__
+
+#include "../feature_management.h"
+
+struct Feat init_feat_hunger();
+
+#endif
